==> analysis/2-bn_analysis.Rmd <==
---
title: "2-bn_analysis"
author: "Bernard"
date: "2020-09-08"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Introduction

This is an example of a Bayesian network, with directed arcs. I only perform the analysis on a subset of provided variables.

### Import libraries

```{r, message=F, warning=F}

rm (list = ls())
# Helper
library (tidyverse)
library (cowplot)
library (ggpubr)
library (flextable)
library (officer)
# BN
library (bnlearn)

# Model
library (caret)

# Feature parallel
library (doParallel)

# Plot
library (Rgraphviz)
library (bnviewer)

# Helper
return_marginal <- function (x) {
  
  x <-  x[-1, , drop = FALSE]
  x <- apply (x, 1, mean)
  
  return (x)
}

```

# Import data

```{r}

dat <- readRDS ("output/dat.RDS")

load ("output/bn_data2.RData")

```

## Collapse variables

```{r}
dat <- dat %>%
  mutate (ppt_medn_base = (ppt_medn_aff_base + ppt_medn_naff_base)/2,
          ppt_uln_base = (ppt_uln_aff_base + ppt_uln_naff_base)/2,
          ppt_radn_base = (ppt_radn_aff_base + ppt_radn_naff_base)/2,
          ppt_neck_base = (ppt_neck_aff_base + ppt_neck_naff_base)/2,
          ppt_cts_base = (ppt_cts_aff_base + ppt_cts_naff_base)/2,
          ppt_ta_base = (ppt_ta_aff_base + ppt_ta_naff_base)/2) %>%
  select (-c(ppt_medn_aff_base:ppt_ta_naff_base)) 
```


### BN analysis

#### Create blacklist

```{r}

var.excl <- c(grep ("groc", names (dat), value = TRUE) ,
              grep ("mean", names (dat), value = TRUE),
              grep ("ppt", names (dat), value = TRUE), 
              #"age", 
              #"pain_years", 
              #"pain_extent", 
              "aff_side" 
              #"emg"
              )
df.bn = as.data.frame (dat)[, !names (dat) %in% var.excl] %>%
  na.omit()

names (df.bn)[grepl ("years", names (df.bn))] <- "duration"
names (df.bn)[grepl ("extent", names (df.bn))] <- "area"
names (df.bn)[grepl ("cts_base", names (df.bn))] <- "ct_base"
names (df.bn)[grepl ("ppt", names (df.bn))] <- str_remove (names (df.bn)[grepl ("ppt", names (df.bn))], "ppt_")
names (df.bn)[grepl ("cts", names (df.bn))] <- str_remove (names (df.bn)[grepl ("cts", names (df.bn))], "cts_")
names (df.bn)[grepl ("worst", names (df.bn))] <- str_remove (names (df.bn)[grepl ("worst", names (df.bn))], "worst_")


df.bn$grp <- factor (df.bn$grp)

rx.var <- "grp"
demo.var = grep("age|duration|emg|area", colnames (df.bn), value = TRUE)
base.var = grep("_base", colnames (df.bn), value = TRUE)
mth1.var = grep("_1m", colnames (df.bn), value = TRUE)
mth3.var = grep("_3m", colnames (df.bn), value = TRUE)
mth6.var = grep("_6m", colnames (df.bn), value = TRUE)
outcome.var = grep("_12m", colnames (df.bn), value = TRUE)

pair_var <- expand.grid(from = names (df.bn),
                        to = names (df.bn)) %>%
  rownames_to_column()

tiers_keep <- pair_var %>%
  filter (!(grepl (paste0(outcome.var, collapse = "|"),from))) %>%
  filter (!(grepl (paste0(rx.var, collapse = "|"),to))) %>%
  filter (!(grepl (paste0(mth6.var, collapse = "|"), from) &
              grepl (paste0(c(demo.var, base.var, mth1.var, mth3.var), collapse = "|"),to))) %>%
  filter (!(grepl (paste0(mth3.var, collapse = "|"), from) &
              grepl (paste0(c(demo.var, base.var, mth1.var), collapse = "|"),to))) %>%
  filter (!(grepl (paste0(mth1.var, collapse = "|"), from) &
            grepl (paste0(c(demo.var, base.var), collapse = "|"),to))) %>%
  filter (!(grepl (paste0(base.var, collapse = "|"), from) &
          grepl (paste0(c(demo.var), collapse = "|"),to))) %>%
  filter (!(grepl (paste0(rx.var, collapse = "|"), from) &
          grepl (paste0(c(demo.var, base.var), collapse = "|"),to)))

bl <- anti_join(pair_var, tiers_keep, by = "rowname")  %>%
  filter (from != to) %>%
  filter (!(grepl ("_12m", from) &
          grepl ("_12m", to))) %>%
  select (from, to)


```

## Explore

```{r}
df.bn.num <- df.bn %>%
  select_if(is.numeric)

M <- cor(df.bn.num)
corrplot::corrplot(M, method = "circle")
```

## Descriptive

```{r fig.height=15}

ord <- names (df.bn) [!grepl ("emg", names (df.bn))]
df_plot <- df.bn %>% 
  select (-emg) %>%
  pivot_longer(cols = -grp,
               names_to = "var",
               values_to = "val") %>%
  mutate (var = factor (var, levels = ord )) %>%
  group_by(grp, var) %>%
  summarize (Mean = mean (val),
             Sd = sd (val)) %>%
  ggplot () +
  geom_point (aes (x = grp, y = Mean), stat = "identity") + 
  geom_errorbar(aes (x = grp, ymin = Mean - Sd, ymax = Mean + Sd), width = 0) + 
  scale_x_discrete(labels=c("1" = "MT", 
                          "2" = "Surgery")) +
  xlab ("Group") +
  ylab ("Scores") +
  facet_wrap(~ var, ncol = 4, scales = "free") + 
  theme_cowplot()

df_plot

```

```{r, eval = FALSE, include = FALSE}
png ("../manuscript_bn/fig1.png", width = 10, height = 7, units = "in", res = 300)
df_plot
dev.off()
```


#### Build the final model using model averaging

```{r, eval = FALSE}
set.seed (123)
boot <- boot.strength(df.bn,
                      R = 200,
                      algorithm = "hc",
                      algorithm.args = list (blacklist = bl))

  

```

#### Get averaged model

```{r}

avg <-  averaged.network(boot, threshold = 0.5)
fit <-  bn.fit (avg, df.bn, method = "mle")
g <- strength.plot(avg, boot, shape = "ellipse", render = FALSE)
graph::nodeRenderInfo(g) = list(fontsize=14)

# Make color


arc_col <- data.frame(arcs = names (edgeRenderInfo(g)$col)) %>%
  separate(arcs, c("parent", "child"), sep = "~")

coef_fit <- coef(fit) %>%
  map_if (is.vector, ~.x[!grepl("Intercept", names (.x))]) %>%
  map_if (is.matrix, return_marginal)  %>%
  unlist ()

coef_fit <- data.frame(arcs = names (coef_fit), coefs = coef_fit) %>%
  separate(arcs, c ("child", "parent"), sep = "[.]") 

new_col <- arc_col %>%
  left_join(coef_fit, by = c("parent", "child")) %>%
  mutate (coefs = replace_na(coefs,88)) %>%
  mutate (col = ifelse (coefs < 0, "red", 
                        ifelse (coefs == 88, "black", "blue"))) 

new_arc_col <- new_col$col
names (new_arc_col) <-   names (edgeRenderInfo(g)$col)

nodeRenderInfo(g)$fill[base.var] = "cyan"
nodeRenderInfo(g)$fill[demo.var] = "cornsilk"
nodeRenderInfo(g)$fill[mth1.var] = "tan1"
nodeRenderInfo(g)$fill[mth3.var] = "gold"
nodeRenderInfo(g)$fill[mth6.var] = "yellow"
nodeRenderInfo(g)$fill[outcome.var] = "tomato"
edgeRenderInfo(g)$col <- new_arc_col
graph::nodeRenderInfo(g) = list(fontsize=14)


renderGraph(g)


```


```{r, eval = FALSE, include = FALSE}

png ("../manuscript_bn/fig2.png", width = 10, height = 7, units = "in", res = 300)
Rgraphviz::renderGraph(g)
dev.off()
```

```{r, fig.height= 20, shinyapp}
viewer(avg,
       bayesianNetwork.width = "100%",
       bayesianNetwork.height = "80vh",
       bayesianNetwork.layout = "layout_with_sugiyama",
       bayesianNetwork.title="Bayesian Network of CTS recovery",
       node.font = list(color = "black", face="Arial", size = 16),
       bayesianNetwork.footer = "Fig. 1 - Layout with Sugiyama"
)
```


#### Performance evaluation using nested cross validation. 

Inner is bootstrap resampling for model averaging. 
Outer is bootstrap resampling k = 25 for performance evaluation.

```{r, eval = FALSE}

set.seed (1245)

flds <- createFolds(1:nrow(df.bn), 
                            k = 10, returnTrain = TRUE)
n_boot = 200

corr.df.list <- list()

for (k in seq_along(flds)) {
  
  train <-  df.bn [flds[[k]], ] %>% as.data.frame()
  test <- df.bn [-flds[[k]], ] %>% as.data.frame()
  
  
  ############
  
  boot2 <- boot.strength(train,
                      R = 200,
                      algorithm = "hc",
                      algorithm.args = list (blacklist = bl))

  #############
  
  avg2 <-  averaged.network(boot2, threshold = 0.5)
  fit2 <-  bn.fit (avg2, train, method = "mle")
  
  num.var <- test %>%
    select_if (is.numeric) %>%
    names ()
  
  corr.df =  structure(numeric(length (num.var)), names = num.var)
  
  for (n in num.var) {
      corr.df[n] = cor(predict(fit2, 
                               data = test, 
                               node = n, 
                               method = "bayes-lw"),
                       test[n])
  }
  
  corr.df.list[[k]] <- corr.df
  

}

corr.df <- bind_cols (corr.df.list) %>%
  apply (1, mean)

names (corr.df) <- num.var

```

```{r, results = "asis"}
corr.df 
```

# Correlation performance table

```{r, eval = FALSE, include = FALSE}

corr.df_ord <- corr.df[grepl ("pain|func|severe", names (corr.df))] 
correlation <- data.frame(Variable = names (corr.df_ord),
                          Value = corr.df_ord %>% round (2)) %>%
  mutate (Strength = ifelse (abs (Value) <= 0.3, "negligible",
                             ifelse (abs(Value) > 0.3 & abs(Value <= 0.5), "low",
                                     ifelse (abs(Value) > 0.5 & abs(Value <= 0.7), "moderate",
                                             ifelse (abs(Value) > 0.7 & abs(Value <= 0.9), "high",
                                                     "very high")))))

ft <- flextable(correlation) %>%
      set_caption(paste0("Correlation between observed and predicted change values")) %>%
      autofit()

my_path <- paste0("../manuscript_bn/table1_corr.docx")

my_doc <- read_docx()  %>% 
  body_add_flextable(ft)

print (my_doc, target = my_path)

```


# Save data

```{r, eval = FALSE, include = FALSE}

save(avg, 
     bl, 
     rx.var,
     base.var,
     demo.var,
     mth1.var,
     mth3.var,
     mth6.var,
     outcome.var,
     boot, 
     corr.df, 
     df.bn, 
     fit,
     file = "output/bn_data2.RData")

```


# Report

##  Influence on group on function at 12 month

```{r}
set.seed (123)

sim = cpdist(fit, nodes = c("grp", "func_12m"), n = 10^4,
               evidence = (TRUE))

mod <- lm(func_12m ~ grp, data = sim)

summary (mod)

confint(mod)

df_plot <- sim %>%
  group_by(grp) %>%
  summarize (Mean = mean (func_12m),
             Sd = sd (func_12m)) %>%
  ggplot () +
  geom_bar (aes (x = grp, y = Mean), stat = "identity") + 
  scale_x_discrete(labels=c("1" = "MT", 
                            "2" = "Surgery")) +
  geom_errorbar(aes (x = grp, ymin = Mean - Sd, ymax = Mean + Sd), width = 0) + 
  xlab ("Group") +
  ylab ("Function 12m") +
  theme_cowplot()

df_plot
```

```{r, eval = FALSE, include = FALSE}
png ("../manuscript_bn/fig3.png", width = 6, height = 5, units = "in", res = 300)
df_plot
dev.off()
```

### What happens when func_1m was unchanged

```{r}
set.seed (123)
avg.mutilated = mutilated(avg, evidence = list(func_1m  = 0))
#strength.plot(avg.mutilated, boot)

fitted.mutilated = bn.fit (avg.mutilated , df.bn, method = "mle")
fitted.mutilated$func_1m = list(coef = c("(Intercept)" = 0), sd = 0)

sim = cpdist(fitted.mutilated , nodes = c("grp", "func_12m"), n = 10^4,
               evidence = (TRUE))

mod <- lm(func_12m ~ grp, data = sim)

summary (mod)

confint(mod)
```


##  Influence on baseline pain on function at 12 month

```{r}
set.seed (123)

sim = cpdist(fit, nodes = c("pain_base", "func_12m"), n = 10^4,
               evidence = (TRUE))

mod <- lm(func_12m ~ pain_base, data = sim)

summary (mod)
confint(mod)

sim %>%
  ggscatter(x = "pain_base", y = "func_12m", add = "reg.line") +
  stat_regline_equation(label.x = 3, label.y = 4)

```

```{r, eval = FALSE, include = FALSE}
png ("../manuscript_bn/fig4.png", width = 6, height = 5, units = "in", res = 300)
sim %>%
  ggscatter(x = "pain_base", y = "func_12m", add = "reg.line") +
  stat_regline_equation(label.x = 3, label.y = 4)
dev.off()
```


### What happens when dep_base was unchanged

```{r}
set.seed (123)
avg.mutilated = mutilated(avg, evidence = list(dep_base  = 0))
#strength.plot(avg.mutilated, boot)

fitted.mutilated = bn.fit (avg.mutilated , df.bn, method = "mle")
fitted.mutilated$dep_base = list(coef = c("(Intercept)" = 0), sd = 0)

sim = cpdist(fitted.mutilated, nodes = c("pain_base", "func_12m"), n = 10^4,
               evidence = (TRUE))

mod <- lm(func_12m ~ pain_base, data = sim)

summary (mod)
confint(mod)

sim %>%
  ggscatter(x = "pain_base", y = "func_12m", add = "reg.line") +
  stat_regline_equation(label.x = 3, label.y = 4)


```

##  Influence on severity at 1month  on function at 12 month

```{r}
set.seed (123)

sim = cpdist(fit, nodes = c("severe_1m", "func_12m"), n = 10^4,
               evidence = (TRUE))

mod <- lm(func_12m ~ severe_1m, data = sim)

summary (mod)
confint(mod)

sim %>%
  ggscatter(x = "severe_1m", y = "func_12m", add = "reg.line") +
  stat_regline_equation(label.x = 3, label.y = 4)

```

```{r, eval = FALSE, include = FALSE}
png ("../manuscript_bn/fig5.png", width = 6, height = 5, units = "in", res = 300)
sim %>%
  ggscatter(x = "severe_1m", y = "func_12m", add = "reg.line") +
  stat_regline_equation(label.x = 3, label.y = 4)
dev.off()
```

### What happens when 6 month function was unchanged

```{r}
set.seed (123)
avg.mutilated = mutilated(avg, evidence = list(func_6m  = 0))
#strength.plot(avg.mutilated, boot)

fitted.mutilated = bn.fit (avg.mutilated , df.bn, method = "mle")
fitted.mutilated$func_6m = list(coef = c("(Intercept)" = 0), sd = 0)

sim = cpdist(fitted.mutilated, nodes = c("severe_1m", "func_12m"), n = 10^4,
               evidence = (TRUE))

mod <- lm(func_12m ~ severe_1m, data = sim)

summary (mod)
confint(mod)

sim %>%
  ggscatter(x = "severe_1m", y = "func_12m", add = "reg.line") +
  stat_regline_equation(label.x = 3, label.y = 4)


```

##  Influence of EMG on function at 12 month

```{r}
set.seed (123)

sim = cpdist(fit, nodes = c("emg", "func_12m"), n = 10^4,
               evidence = (TRUE))

mod <- lm(func_12m ~ emg, data = sim)

summary (aov(mod))
TukeyHSD(aov(mod))

df_plot <- sim %>%
  group_by(emg) %>%
  summarize (Mean = mean (func_12m),
             Sd = sd (func_12m)) %>%
  ggplot () +
  geom_bar (aes (x = emg, y = Mean), stat = "identity") + 
  scale_x_discrete(labels=c("1" = "Mild", 
                            "2" = "Moderate",
                            "3" = "Severe")) +
  geom_errorbar(aes (x = emg, ymin = Mean - Sd, ymax = Mean + Sd), width = 0) + 
  xlab ("EMG classification") +
  ylab ("Function 12m") +
  theme_cowplot()

df_plot
```

```{r, eval = FALSE, include = FALSE}
png ("../manuscript_bn/fig6.png", width = 6, height = 5, units = "in", res = 300)
df_plot
dev.off()
```
