==> analysis/3-network.Rmd <==
---
title: "3-network"
author: "bernard-liew"
date: "2020-09-24"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

# Load package

```{r}
if (!require("pacman")) install.packages("pacman")
pacman::p_load(tidyverse,
               qgraph,
               stats,
               bootnet,
               igraph,
               arsenal,
               flextable,
               officer,
               cowplot,
               huge)
```

# Import data

```{r}

if (file.exists("output/network_res.RDS")) {
  
  dat2 <- readRDS("output/network_res.RDS")
  
  list2env(dat2 ,.GlobalEnv)
  
  skip_eval <- TRUE

  } else {
  
  dat <- readRDS ("output/dat.RDS")
  
  skip_eval <- FALSE
  }

# Helper for nice descriptive table

meanNsd_transform <- function (x) {

  m <- round (x[[1]][1], 2)
  s <- round (x[[1]][2], 2)

  m_s <- paste0(m, "(", s, ")")

  return (m_s)
}

new_labs <- c("Pain intensity",
              "PPT median nerve",
              "PPT ulnar nerve",
              "PPT radial nerve",
              "PPT neck",
              "PPT carpal tunnel",
              "PPT tibialis anterior",
              "Function",
              "Symptom severity",
              "Depressive levels")

```

# Clean data

```{r}

if (skip_eval == FALSE) {
  
  dat_base <- dat[, !grepl ("naff.*base", names (dat))] 
  dat_base <- dat_base[, grepl ("aff|base", names (dat_base))] 
  dat_base <- dat_base %>%
    select (-worst_pain_base, - aff_side) %>%
    rename (func = cts_func_base,
            severity = cts_severe_base) 
  names (dat_base) <- str_remove_all (names(dat_base), "_base|mean_|_aff")
  
  node_labs <- names(dat_base)
  
  names(dat_base) <- paste0("V", 1:ncol(dat_base))
  
  dat_base %>%
    pivot_longer(cols = everything (),
                 names_to = "var",
                 values_to = "val") %>%
    ggplot () +
    geom_histogram (aes (val)) +
    facet_wrap (~ var, ncol = 5, scales = "free_x")
  
}
```

# Descriptive summary

```{r, results = "asis" }

dat_sub <- dat %>%
  dplyr::select (age, 
                 pain_years, 
                 mean_pain_base,
                 ppt_medn_aff_base, 
                 ppt_uln_aff_base,
                 ppt_radn_aff_base,
                 ppt_neck_aff_base,
                 ppt_cts_aff_base,
                 ppt_ta_aff_base,
                 cts_func_base,
                 cts_severe_base,
                 dep_base)

names (dat_sub) <- str_remove_all (names(dat_sub), "_base|mean_|_aff")

tab1 <- tableby ( ~. , data = dat_sub, digits = 2, digits.p = 2) 

summary (tab1)
```

```{r eval = FALSE, include = FALSE}
tab1 <- tab1 %>%
  as.data.frame() %>%
  filter (!term %in% c("Nmiss", "range")) %>%
  select (-c(group.term:term, variable.type,Total)) 

tab2 <-  tab1[-seq (1, 24, 2),]
tab2$label <- tab1[seq (1, 24, 2),1]

for(row in 1:nrow(tab2)) {
  for(col in 2:ncol(tab2)) {
    tab2[row, col] <- meanNsd_transform (tab2[row, col])
  }
}

colnames (tab2) <- c ("Baseline variable", "Mean (SD)")

my_path <- paste0("../manuscript/table_1_", 
                  "baseline",
                  ".docx")

ft <- flextable(tab2) %>%
  autofit()

my_doc <- read_docx()  %>% 
  body_add_flextable(ft) %>%
  body_end_section_landscape()

print (my_doc, target = my_path)
```



# Network analysis

## On baseline data

```{r}

if (skip_eval == FALSE) {
  
  df <- dat_base
  df <- huge.npn (df)

  set.seed (1)
  nw <- estimateNetwork(df, 
                      default="EBICglasso",
                      corMethod = "cor",
                      tuning = 0.5,
                      lambda.min.ratio = 0.001,
                      corArgs =
                        list(method = "pearson",
                             use = "pairwise.complete.obs"))
  
}



```

## Plot network

Arcs in blue means a positive correlation between connecting variables.

Arcs in red means a netative correlation between connecting variables.

Thickness of arcs gives you a qualitative indication of correlation magnitude.


```{r}
nw_plot <- plot (nw, nodeNames = node_labs,  layout = "spring", legend = TRUE)

wts_df <- nw_plot$Edgelist %>%
  bind_cols() %>%
  select (from, to, weight) %>%
  mutate_all (round, 2)

wts_df %>% 
  mutate (weight_abs = abs (weight)) %>%
  slice_max (weight_abs, n = 10)

wts_df %>% 
  mutate (weight_abs = abs (weight)) %>%
  slice_min (weight_abs, n = 10)
```

```{r eval = FALSE, include = FALSE}
tiff(width = 15, height = 15, units = "in", res = 100, file = "../manuscript/fig_1.tiff")
  plot (nw, nodeNames = new_labs,  layout = "spring", legend = TRUE, legend.cex = 1.2)
dev.off()
```

## Centrality

High centrality nodes have strong connections to many other nodes, and act as hubs that connect otherwise disparate nodes to one another. 

Low centrality nodes exist on the periphery of the network, with fewer and weaker connections to other nodes of the
network.

Strength is the sum of the absolute value of its connections with other nodes in the network.

Closeness centrality is defined as the inverse of the sum of the distances of the focal node from all the other nodes in the network. Closeness is the average shortest path between a given node and the remaining nodes in the network. Nodes with higher closeness are more proximally connected to the rest of the network.

Betweenness is the number of times in which a given node lies on the shortest path between two other nodes.

The greater the value of centrality indices to one, the more important the variable.

```{r}

centr <- centralityPlot(nw,  
                        include = c("Strength", "Closeness", "Betweenness"), 
                        scale = "relative", 
                        labels = new_labs)

```

```{r eval = FALSE, include = FALSE}
tiff(width = 10, height = 8, units = "in", res = 200, file = "../manuscript/fig_3.tiff")
  centr + 
    theme(text = element_text(size = 14)) +
    xlim (0, 1)
dev.off()
```

## Accuracy

### Edge weights stability

```{r message=FALSE, warning=FALSE}
if (skip_eval == FALSE) {
  edge_wts <- bootnet(nw,
                    nBoots = 2000,
                    nCores = 6,
                    statistics = "edge")
 
}

plot (edge_wts, satistics = "edge", plot = "area", order = "sample", CIstyle = "SE")

```

```{r eval = FALSE, include = FALSE}
tiff(width = 8, height = 15, units = "in", res = 200, file = "../manuscript/fig_2.tiff")
  plot (edge_wts, satistics = "edge", plot = "area", order = "sample", CIstyle = "SE") + 
    theme(text = element_text(size = 16)) 
dev.off()
```


```{r eval = FALSE, include = FALSE}
p_val <- plot (edge_wts, plot = "difference", satistics = "edge", order = "sample", onlyNonZero = TRUE, verbose = TRUE)
p_val

```


### Centrality stability

```{r}
if (skip_eval == FALSE) {
  stats2boot <- c("edge", "strength", "expectedInfluence", "closeness", "betweenness")
  centr_stb <- bootnet(nw,
                       nBoots = 2000,
                       nCores = 6,
                       statistics = stats2boot,
                       type = "case")
  
  cor_stb <-  corStability (centr_stb)

}

plot (centr_stb, statistics = c("strength", "closeness", "betweenness"))
```

```{r eval = FALSE, include = FALSE}
tiff(width = 5, height = 5, units = "in", res = 100, file = "../manuscript/fig_4.tiff")
  plot (centr_stb, statistics = c("strength", "closeness", "betweenness")) + 
    theme(text = element_text(size = 14)) 
dev.off()
```

```{r eval = FALSE, include = FALSE}
Sim <- netSimulator(
  input = nw$graph,
  default = "EBICglasso",
  nCases = c(100,250,500,1000,2500),
  tuning = 0.5,
  nReps = 1000,
  nCores = 8
)


plot (Sim)

plot(Sim5,
     yvar = c("strength","closeness","betweenness"))
```

# Save data

```{r, eval = FALSE}
dat2save <- list ("dat" = dat,
                  "df" = df,
                  "nw" = nw,
                  "nw_plot" = nw_plot,
                  "centr" = centr,
                  "centr_stb" = centr_stb,
                  "edge_wts" = edge_wts,
                  "cor_stb" = cor_stb,
                  "node_labs" = node_labs,
                  "stats2boot" = stats2boot,
                  "Sim" = Sim)

saveRDS(dat2save,
        "output/network_res.RDS")
```
